Fix clippy warnings

+++ glean-core/ffi/examples/glean.h
@@ -12,7 +12,7 @@ uint8_t glean_is_upload_enabled();
 uint64_t glean_new_boolean_metric(const char* name, const char* category, struct ExternError *err);
 
 uint64_t glean_new_counter_metric(const char* name, const char* category, struct ExternError *err);
-void glean_counter_add(uint64_t handle, uint32_t amount, struct ExternError *err);
+void glean_counter_add(uint64_t handle, uint64_t amount, struct ExternError *err);
 
 uint64_t glean_new_string_metric(const char* name, const char* category, struct ExternError *err);
